#33: Move webservice to separate package
