+++ backend/src/main/kotlin/xyz/elitese/ehrenamtskarte/webservice/Entrypoint.kt
@@ -1,7 +1,7 @@
 /*
  * This Kotlin source file was generated by the Gradle 'init' task.
  */
-package xyz.elitese.ehrenamtskarte
+package xyz.elitese.ehrenamtskarte.webservice
 
 import io.javalin.Javalin
 import io.javalin.http.staticfiles.Location
