+++ backend/src/main/kotlin/xyz/elitese/ehrenamtskarte/webservice/dataloader/CategoryDataLoader.kt
@@ -1,8 +1,8 @@
-package xyz.elitese.ehrenamtskarte.dataloader
+package xyz.elitese.ehrenamtskarte.webservice.dataloader
 
 import kotlinx.coroutines.runBlocking
 import org.dataloader.DataLoader
-import xyz.elitese.ehrenamtskarte.schema.types.Category
+import xyz.elitese.ehrenamtskarte.webservice.schema.types.Category
 import java.util.concurrent.CompletableFuture
 
 const val CATEGORY_LOADER_NAME = "CATEGORY_LOADER"
