+++ backend/src/main/kotlin/xyz/elitese/ehrenamtskarte/webservice/dataloader/initalizeDataLoaderRegistry.kt
@@ -1,6 +1,5 @@
-
 import org.dataloader.DataLoaderRegistry
-import xyz.elitese.ehrenamtskarte.dataloader.*
+import xyz.elitese.ehrenamtskarte.webservice.dataloader.*
 
 fun initializeDataLoaderRegistry(dataLoaderRegistry: DataLoaderRegistry) {
     dataLoaderRegistry.register(CONTACT_LOADER_NAME, batchContactLoader)
