+++ backend/src/main/kotlin/xyz/elitese/ehrenamtskarte/webservice/schema/types/Category.kt
@@ -1,4 +1,4 @@
-package xyz.elitese.ehrenamtskarte.schema.types
+package xyz.elitese.ehrenamtskarte.webservice.schema.types
 
 data class Category(
         val id: Long,
