+++ backend/src/main/kotlin/xyz/elitese/ehrenamtskarte/webservice/schema/types/Contact.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package xyz.elitese.ehrenamtskarte.schema.types
+package xyz.elitese.ehrenamtskarte.webservice.schema.types
 
 
 data class Contact(
